+++ serpentTools/__init__.py
@@ -2,6 +2,7 @@
 from serpentTools.parsers import *
 from serpentTools.data import *
 from serpentTools.samplers import *
+from serpentTools.seed import *
 from ._version import get_versions
 __version__ = get_versions()['version']
 del get_versions